+++ frontend/src/components/navbar/nav.ts
@@ -1,0 +1,11 @@
+const navbarItem: NavbarLinkProps[] = [
+    {
+      name: "Home",
+      path: "/",
+    },
+    {
+      name: "Product",
+      path: "/product",
+    },
+  ];
+
